adjust incoming call ui to figma
