+++ ui/src/main.rs
@@ -5,13 +5,11 @@ use clap::Parser;
 use common::icons::outline::Shape as Icon;
 use common::icons::Icon as IconElement;
 use common::language::get_local_text;
-use common::warp_runner::BlinkCmd;
 
 use common::notifications::{NotificationAction, NOTIFICATION_LISTENER};
 use common::warp_runner::ui_adapter::MessageEvent;
 use common::warp_runner::WarpEvent;
 use common::{get_extras_dir, warp_runner, LogProfile, STATIC_ARGS, WARP_CMD_CH, WARP_EVENT_CH};
-use components::calldialog::CallDialog;
 use dioxus::prelude::*;
 use dioxus_desktop::tao::dpi::LogicalSize;
 use dioxus_desktop::tao::event::WindowEvent;
@@ -24,8 +22,6 @@ use futures::StreamExt;
 use kit::components::context_menu::{ContextItem, ContextMenu};
 use kit::components::nav::Route as UIRoute;
 use kit::components::topbar_controls::Topbar_Controls;
-use kit::components::user_image::UserImage;
-use kit::components::user_image_group::UserImageGroup;
 use kit::elements::button::Button;
 use kit::elements::Appearance;
 use kit::layout::modal::Modal;
@@ -33,8 +29,6 @@ use layouts::friends::FriendRoute;
 use notify::{RecommendedWatcher, RecursiveMode, Watcher};
 use once_cell::sync::Lazy;
 use overlay::{make_config, OverlayDom};
-use utils::build_user_from_identity;
-use uuid::Uuid;
 
 use std::collections::HashMap;
 
@@ -68,7 +62,6 @@ use crate::utils::auto_updater::{
     DownloadProgress, DownloadState, SoftwareDownloadCmd, SoftwareUpdateCmd,
 };
 
-use crate::utils::build_participants;
 use crate::utils::get_drag_event::DRAG_EVENT;
 use crate::window_manager::WindowManagerCmdChannels;
 use crate::{components::chat::RouteInfo, layouts::chat::ChatLayout};
@@ -518,7 +511,6 @@ fn app(cx: Scope) -> Element {
                 id: "app-wrap",
                 get_titlebar{},
                 get_toasts{},
-                get_call_dialog{},
                 get_router{},
                 get_logger{},
             },
@@ -1163,104 +1155,6 @@ fn get_titlebar(cx: Scope) -> Element {
             })),
         },
     ))
-}
-
-enum CallDialogCmd {
-    Accept(Uuid),
-    Reject(Uuid),
-}
-
-fn get_call_dialog(cx: Scope) -> Element {
-    let state = use_shared_state::<State>(cx)?;
-    let ch = use_coroutine(cx, |mut rx| {
-        to_owned![state];
-        async move {
-            let warp_cmd_tx = WARP_CMD_CH.tx.clone();
-            while let Some(cmd) = rx.next().await {
-                match cmd {
-                    CallDialogCmd::Accept(id) => {
-                        let (tx, rx) = oneshot::channel();
-                        if let Err(_e) = warp_cmd_tx.send(WarpCmd::Blink(BlinkCmd::AnswerCall {
-                            call_id: id,
-                            rsp: tx,
-                        })) {
-                            log::error!("failed to send blink command");
-                            continue;
-                        }
-
-                        match rx.await {
-                            Ok(_) => {
-                                state.write().mutate(Action::AnswerCall(id));
-                            }
-                            Err(e) => {
-                                log::error!("warp_runner failed to answer call: {e}");
-                            }
-                        }
-                    }
-                    CallDialogCmd::Reject(id) => {
-                        let (tx, rx) = oneshot::channel();
-                        if let Err(_e) = warp_cmd_tx.send(WarpCmd::Blink(BlinkCmd::RejectCall {
-                            call_id: id,
-                            rsp: tx,
-                        })) {
-                            log::error!("failed to send blink command");
-                            continue;
-                        }
-
-                        match rx.await {
-                            Ok(_) => {
-                                state.write().ui.call_info.reject_call(id);
-                            }
-                            Err(e) => {
-                                log::error!("warp_runner failed to answer call: {e}");
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    });
-
-    let call = match state.read().ui.call_info.active_call() {
-        Some(_) => return None,
-        None => match state.read().ui.call_info.pending_calls().first() {
-            Some(call) => call.clone(),
-            None => return None,
-        },
-    };
-    let mut participants = state.read().get_identities(&call.participants);
-    let own_id = state.read().did_key();
-    participants.retain(|x| x.did_key() != own_id);
-
-    let my_identity = build_user_from_identity(state.read().get_own_identity());
-
-    cx.render(rsx!(CallDialog {
-        caller: cx.render(rsx!(UserImageGroup {
-            participants: build_participants(&participants),
-            with_username: State::join_usernames(&participants),
-        },)),
-        callee: cx.render(rsx!(UserImage {
-            platform: my_identity.platform,
-            status: my_identity.status,
-            image: my_identity.photo,
-            with_username: my_identity.username,
-        })),
-        description: get_local_text("remote-controls.incoming-call"),
-        with_accept_btn: cx.render(rsx!(Button {
-            icon: Icon::Phone,
-            appearance: Appearance::Success,
-            onpress: move |_| {
-                ch.send(CallDialogCmd::Accept(call.id));
-            }
-        })),
-        with_deny_btn: cx.render(rsx!(Button {
-            icon: Icon::PhoneXMark,
-            appearance: Appearance::Danger,
-            onpress: move |_| {
-                ch.send(CallDialogCmd::Reject(call.id));
-            }
-        })),
-    }))
 }
 
 fn get_router(cx: Scope) -> Element {
